Attempt at testing patterns to autosync the video blanking to a timer
Using 50/60Hz attributes in strategic places and times.

Python script to test hypotheses, and Gnuplot script to visualize data.

The search space consists of all possible lines the ULA is currently on
(including the blanking ones) when we start the timer, and the current
frequency (rather max lines) for this possibility.

Then for each scanline we calculate what would happen, and how the delta
(how many lines we lag).

The objective is to minimize the delta discrepancy (ie. for each possible
initial condition, be as close possible to the blanking we want to sync to).

Currently we get to 51 lines maximum spread.